# bandcoord/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'BandCoord API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://www.iestrassierra.net/alumnado/curso2425/DAW/daw2425a16/laravel'
...
paths:
...
  /api/mailTo:
    post:
      summary: 'Enviar correo personalizado'
      operationId: enviarCorreoPersonalizado
      description: 'Envía un correo electrónico personalizado a un usuario.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: 'Correo enviado correctamente con vista personalizada'
                properties:
                  success:
                    type: string
                    example: 'Correo enviado correctamente con vista personalizada'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Datos inválidos'
                  messages:
                    email:
                      - 'El campo email es obligatorio.'
                    mensaje:
                      - 'El campo mensaje es obligatorio.'
                properties:
                  error:
                    type: string
                    example: 'Datos inválidos'
                  messages:
                    type: object
                    properties:
                      email:
                        type: array
                        example:
                          - 'El campo email es obligatorio.'
                        items:
                          type: string
                      mensaje:
                        type: array
                        example:
                          - 'El campo mensaje es obligatorio.'
                        items:
                          type: string
      tags:
        - 'Gestión de Usuarios'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Email del destinatario.'
                  example: usuario@ejemplo.com
                  nullable: false
                mensaje:
                  type: string
                  description: 'Contenido del mensaje (máximo 1000 caracteres).'
                  example: 'Este es un mensaje personalizado.'
                  nullable: false
              required:
                - email
                - mensaje
      security: []
...